refactor: extract class
